worker/requirements.txt
aio-pika==9.4.1
redis==5.0.4
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic-settings==2.3.3
python-dotenv==1.0.1